Moved ArrayIterator inside helper package
